Remove obsolete navigation e2e tests
