+++ integration-tests/tests/pages/navigation.page.ts
@@ -1,19 +1,17 @@
-import type { Page } from '@playwright/test';
+import { Page, expect } from '@playwright/test';
 
-import { expect } from '../e2e.fixture.ts';
-
-export class NavigationFixture {
+export class NavigationPage {
     constructor(public readonly page: Page) {}
 
-    public async clickLink(name: string) {
+    async clickLink(name: string) {
         await this.page.getByRole('link', { name, exact: true }).click();
     }
 
-    public async openOrganismNavigation() {
+    async openOrganismNavigation() {
         await this.page.getByText('Organisms').click();
     }
 
-    public async expectTitle(title: string) {
+    async expectTitle(title: string) {
         await expect(this.page).toHaveTitle(new RegExp(`^${title}`));
     }
 }
